Added iboutlets a few TVC’s.  need to finish connecting the rest.

+++ 90 DWT BB/Build_Back+Bi_1_TVC.m
@@ -490,7 +490,7 @@
 {
     
     NSString *headerTitle = @"";
-    headerTitle = [headerTitle stringByAppendingFormat:@"Set %d of %d", section + 1, self.tableView.numberOfSections];
+    headerTitle = [headerTitle stringByAppendingFormat:@"Set %ld of %ld", section + 1, (long)self.tableView.numberOfSections];
     
     return headerTitle;
 }
